Added base class for TranslationProviders
